requirements-dev.txt
django>=2.0
invoke
cherrypy<=11.0.0
pymysql
packaging
django-formtools
django-recaptcha
docutils
pyyaml
passlib
Pillow
raven
django-htmlmin
celery
redis
django-redis
flower
urllib3<1.23,>=1.21.1
pip>=10
pycryptodome
colorlog
markdown
aenum
graphviz
# In here you can specify pip modules that should be installed
dnspython
requests
django-oauth-toolkit
django-cors-middleware
requests_oauthlib>=0.4.2
oauthlib>=2.0.6
ldap3
beautifulsoup4>=4.6.0
lxml>=3.8.0
signxml
defusedxml
slumber
apache-libcloud
slumber
django-two-factor-auth
django-revproxy
influxdb>=4.1.1
psutil
requests
coverage
isort
pylint<2.0.0
astroid<2.0,>=1.6
pylint-django<=0.9.4
prospector<1.0
django-debug-toolbar
recommonmark
autopep8
pycodestyle<2.4.0,>=2.0.0
-e git+https://github.com/BeryJu/pdoc.git@ebb38ec5d2a023a31442757bba697c60d315578f#egg=pdoc
markdown
mkdocs
mkdocs-material
tornado>=5.0
pyroma<2.4
safety
bandit<1.5.0
unittest-xml-reporting
